ai_service/requirements.txt
flask==3.0.0
tensorflow==2.15.0
numpy==1.26.2
pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
